+++ cedar-policy/src/api.rs
@@ -1701,9 +1701,10 @@ impl<'a> Diagnostic for ValidationWarning<'a> {
 
 /// Identifier portion of the [`EntityUid`] type.
 ///
-/// An `EntityId` can can be constructed using [`EntityId::from_str`] or by
-/// calling `parse()` on a string. This implementation is `Infallible`, so the
-/// parsed `EntityId` can be extracted safely.
+/// All strings are valid [`EntityId`]s, and can be
+/// constructed either using [`EntityId::new`]
+/// or by using the implementation of [`FromStr`]. This implementation is [`Infallible`], so the
+/// parsed [`EntityId`] can be extracted safely.
 ///
 /// ```
 /// # use cedar_policy::EntityId;
@@ -1713,6 +1714,16 @@ impl<'a> Diagnostic for ValidationWarning<'a> {
 #[repr(transparent)]
 #[derive(Debug, Clone, Hash, PartialEq, Eq, PartialOrd, Ord, RefCast)]
 pub struct EntityId(ast::Eid);
+
+impl EntityId {
+    /// Construct an [`EntityId`] from a source string
+    pub fn new(src: impl AsRef<str>) -> Self {
+        match src.as_ref().parse() {
+            Ok(eid) => eid,
+            Err(infallible) => match infallible {},
+        }
+    }
+}
 
 impl FromStr for EntityId {
     type Err = Infallible;
@@ -2680,20 +2691,29 @@ impl TemplateResourceConstraint {
 
 /// Unique ids assigned to policies and templates.
 ///
-/// A `PolicyId` can can be constructed using [`PolicyId::from_str`] or by
-/// calling `parse()` on a string. This currently always returns `Ok()`.
-///
+/// A [`PolicyId`] can can be constructed using [`PolicyId::from_str`] or by
+/// calling `parse()` on a string.
+/// This implementation is [`Infallible`], so the parsed [`EntityId`] can be extracted safely.
+/// Examples:
 /// ```
 /// # use cedar_policy::PolicyId;
-/// let id : PolicyId = "my-id".parse().unwrap();
+/// let id = PolicyId::new("my-id");
+/// let id : PolicyId = "my-id".parse().unwrap_or_else(|never| match never {});
 /// # assert_eq!(id.as_ref(), "my-id");
 /// ```
 #[repr(transparent)]
 #[derive(Debug, PartialEq, Eq, Clone, Hash, Serialize, Deserialize, RefCast)]
 pub struct PolicyId(ast::PolicyID);
 
+impl PolicyId {
+    /// Construct a [`PolicyId`] from a source string
+    pub fn new(id: impl AsRef<str>) -> Self {
+        Self(ast::PolicyID::from_string(id.as_ref()))
+    }
+}
+
 impl FromStr for PolicyId {
-    type Err = ParseErrors;
+    type Err = Infallible;
 
     /// Create a `PolicyId` from a string. Currently always returns `Ok()`.
     fn from_str(id: &str) -> Result<Self, Self::Err> {